compress/requirements.txt
numpy==1.26.4
Pillow==10.3.0
scikit_learn==1.4.1.post1
